[restructuring for better dispatch]
This commit moved a lot of things around basically just to make the ExchangeManager no longer a TypedActor so that it could use a router.

I realized that the no-router TypedActor sitting there would be an extreme chokepoint when dispatching large ammounts of messages because
they would all be going through the same instance.  Now we can use smallest mailbox, round-robin, gather-scatter, or any of the other
router patterns to back ExchangeManager.  This should make it more performant in the long run.

[testing!]

Added skeleton for test classes and eventual test layout