[restructuring]
removed a lot of the hibernate data stuff because that isn't really tennable for right now.

restructured the achoo management stuff so that it gets started directly and doesn't
have to go through a management object.  however, the .stop() method can't be handled
by jetty without using akka for dispatch because if you do that you get interupted
exceptions either while waiting for the blocking .stop() *or* while using a plain
runnable to cause the stop.  [removed two files]