[rearchitecting]
Using transactors instead of hacked up mechanism.  Transactors increment or otherwise atomically modify the subscriber list.  This prevents running out
of subscribers for no apparent reason.  The getSubscribers() action is not atomic... so that means that it won't hose up high throughput.

[minor]

Added versions to the config file so that we can show that information in the start-up script